requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
jinja2==3.1.2
python-multipart==0.0.6
